# Conflicts:
#	scripts/build.py

+++ engine/physics/src/physics/physics_3d.cpp
@@ -1,12 +1,12 @@
-// Copyright 2020-2022 The Defold Foundation
+// Copyright 2020-2025 The Defold Foundation
 // Copyright 2014-2020 King
 // Copyright 2009-2014 Ragnar Svensson, Christian Murray
 // Licensed under the Defold License version 1.0 (the "License"); you may not use
 // this file except in compliance with the License.
-// 
+//
 // You may obtain a copy of the License, together with FAQs at
 // https://www.defold.com/license
-// 
+//
 // Unless required by applicable law or agreed to in writing, software distributed
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
@@ -143,8 +143,7 @@ namespace dmPhysics
         ToBt(params.m_WorldMin, world_aabb_min, context->m_Scale);
         btVector3 world_aabb_max;
         ToBt(params.m_WorldMax, world_aabb_max, context->m_Scale);
-        int maxProxies = 1024;
-        m_OverlappingPairCache = new btAxisSweep3(world_aabb_min,world_aabb_max,maxProxies);
+        m_OverlappingPairCache = new btAxisSweep3(world_aabb_min,world_aabb_max, params.m_MaxCollisionObjectsCount);
 
         m_Solver = new btSequentialImpulseConstraintSolver;
 
@@ -166,6 +165,11 @@ namespace dmPhysics
         delete m_OverlappingPairCache;
         delete m_Dispatcher;
         delete m_CollisionConfiguration;
+    }
+
+    HContext3D GetContext3D(HWorld3D world)
+    {
+        return world->m_Context;
     }
 
     static void ResponseFromRayCastResult(RayCastResponse& response, btScalar inv_scale, btScalar fraction,
@@ -288,7 +292,7 @@ namespace dmPhysics
         context->m_TriggerOverlapCapacity = params.m_TriggerOverlapCapacity;
         context->m_AllowDynamicTransforms = params.m_AllowDynamicTransforms;
         dmMessage::Result result = dmMessage::NewSocket(PHYSICS_SOCKET_NAME, &context->m_Socket);
-        if (result != dmMessage::RESULT_OK)
+        if (result != dmMessage::RESULT_OK && result != dmMessage::RESULT_SOCKET_EXISTS)
         {
             dmLogFatal("Could not create socket '%s'.", PHYSICS_SOCKET_NAME);
             DeleteContext3D(context);
@@ -369,7 +373,7 @@ namespace dmPhysics
         // Update all trigger transforms before physics world step
         if (world->m_GetWorldTransform != 0x0)
         {
-            DM_PROFILE(Physics, "UpdateTriggers");
+            DM_PROFILE("UpdateTriggers");
             int collision_object_count = world->m_DynamicsWorld->getNumCollisionObjects();
             btCollisionObjectArray& collision_objects = world->m_DynamicsWorld->getCollisionObjectArray();
             for (int i = 0; i < collision_object_count; ++i)
@@ -421,7 +425,7 @@ namespace dmPhysics
         }
 
         {
-            DM_PROFILE(Physics, "StepSimulation");
+            DM_PROFILE("StepSimulation");
 
             // Default behavior for Bullet3D is to use fixed timesteps with max_steps=1 and fixed_timestep=1.0f/60.0f
             float dt = step_context.m_DT;
@@ -435,7 +439,7 @@ namespace dmPhysics
         uint32_t size = world->m_RayCastRequests.Size();
         if (size > 0)
         {
-            DM_PROFILE(Physics, "RayCasts");
+            DM_PROFILE("RayCasts");
             for (uint32_t i = 0; i < size; ++i)
             {
                 const RayCastRequest& request = world->m_RayCastRequests[i];
@@ -477,7 +481,7 @@ namespace dmPhysics
         float contact_impulse_limit = world->m_Context->m_ContactImpulseLimit;
         if (collision_callback != 0x0 || contact_point_callback != 0x0)
         {
-            DM_PROFILE(Physics, "CollisionCallbacks");
+            DM_PROFILE("CollisionCallbacks");
             int num_manifolds = dispatcher->getNumManifolds();
             for (int i = 0; i < num_manifolds && (requests_collision_callbacks || requests_contact_callbacks); ++i)
             {
@@ -556,7 +560,7 @@ namespace dmPhysics
 
     void UpdateOverlapCache(OverlapCache* cache, HContext3D context, btDispatcher* dispatcher, const StepWorldContext& step_context)
     {
-        DM_PROFILE(Physics, "TriggerCallbacks");
+        DM_PROFILE("TriggerCallbacks");
         OverlapCacheReset(cache);
         OverlapCacheAddData add_data;
         add_data.m_TriggerEnteredCallback = step_context.m_TriggerEnteredCallback;
@@ -694,10 +698,7 @@ namespace dmPhysics
         float object_scale = 1.0f;
         if (has_world_transform)
         {
-            if( data.m_Type != COLLISION_OBJECT_TYPE_TRIGGER)
-            {
-                object_scale = world_transform.GetUniformScale();
-            }
+            object_scale = world_transform.GetUniformScale();
         }
 
         bool clone_shapes = world->m_AllowDynamicTransforms || object_scale != 1.0f;
@@ -817,6 +818,25 @@ namespace dmPhysics
         world->m_DynamicsWorld->removeCollisionObject(bt_co);
         delete bt_co;
         delete co;
+    }
+
+    HCollisionShape3D GetCollisionShape3D(HCollisionObject3D collision_object, uint32_t index)
+    {
+        btCollisionShape* shape = GetCollisionObject(collision_object)->getCollisionShape();
+
+        if (shape->isCompound())
+        {
+            btCompoundShape* compound = (btCompoundShape*)shape;
+            if (compound->getNumChildShapes() > index)
+            {
+                return compound->getChildShape(index);
+            }
+        }
+        else if (index == 0)
+        {
+            return shape;
+        }
+        return 0;
     }
 
     uint32_t GetCollisionShapes3D(HCollisionObject3D collision_object, HCollisionShape3D* out_buffer, uint32_t buffer_size)
@@ -959,7 +979,7 @@ namespace dmPhysics
 
     void SetEnabled3D(HWorld3D world, HCollisionObject3D collision_object, bool enabled)
     {
-        DM_PROFILE(Physics, "SetEnabled");
+        DM_PROFILE("SetEnabled3D");
         bool prev_enabled = IsEnabled3D(collision_object);
         // avoid multiple adds/removes
         if (prev_enabled == enabled)
@@ -1013,7 +1033,7 @@ namespace dmPhysics
         btCollisionObject* co = GetCollisionObject(collision_object);
         return co->getActivationState() == ISLAND_SLEEPING;
     }
-    
+
     void Wakeup3D(HCollisionObject3D collision_object)
     {
         btCollisionObject* co = GetCollisionObject(collision_object);
@@ -1092,10 +1112,10 @@ namespace dmPhysics
     void SetGroup3D(HWorld3D world, HCollisionObject3D collision_object, uint16_t groupbit) {
 		CollisionObject3D* co = (CollisionObject3D*) collision_object;
 		btCollisionObject* bt_co = co->m_CollisionObject;
-		
+
 		bool enabled = IsEnabled3D(collision_object);
-		
-		if (!enabled) { 
+
+		if (!enabled) {
 			// the collision object off the world. We just have to update the group property in CollisionObject3D. When the co is enabled, the group property will get applied.
 			co->m_CollisionGroup = groupbit;
 		} else {
@@ -1116,23 +1136,23 @@ namespace dmPhysics
 	bool GetMaskBit3D(HCollisionObject3D collision_object, uint16_t groupbit) {
 		CollisionObject3D* co = (CollisionObject3D*) collision_object;
 		uint16_t maskbits = co->m_CollisionMask;
-		return !!(maskbits & groupbit);		
+		return !!(maskbits & groupbit);
 	}
-	
+
     void SetMaskBit3D(HWorld3D world, HCollisionObject3D collision_object, uint16_t groupbit, bool boolvalue) {
 		CollisionObject3D* co = (CollisionObject3D*) collision_object;
 		btCollisionObject* bt_co = co->m_CollisionObject;
-		
+
 		bool enabled = IsEnabled3D(collision_object);
-		
+
 		//calculate new mask once
 		uint16_t newmask = co->m_CollisionMask;
 		if (boolvalue)
 			newmask |= groupbit;
 		else
 			newmask &= ~groupbit;
-		
-		if (!enabled) { 
+
+		if (!enabled) {
 			// the collision object off the world. We just have to update the mask property in CollisionObject3D. When the co is enabled, the mask property will get applied.
 			co->m_CollisionMask = newmask;
 		} else {
@@ -1148,8 +1168,8 @@ namespace dmPhysics
 				world->m_DynamicsWorld->addCollisionObject(bt_co, co->m_CollisionGroup, co->m_CollisionMask);
 			}
 		}
-	}	
-    
+	}
+
     void RequestRayCast3D(HWorld3D world, const RayCastRequest& request)
     {
         if (!world->m_RayCastRequests.Full())
@@ -1180,7 +1200,7 @@ namespace dmPhysics
 
     void RayCast3D(HWorld3D world, const RayCastRequest& request, dmArray<RayCastResponse>& results)
     {
-        DM_PROFILE(Physics, "RayCasts");
+        DM_PROFILE("RayCast3D");
 
         if (lengthSqr(request.m_To - request.m_From) <= 0.0f)
         {
@@ -1250,9 +1270,70 @@ namespace dmPhysics
         return gravity;
     }
 
+    void GetCollisionShapeRadius3D(HCollisionShape3D shape, float* radius)
+    {
+        btCollisionShape* bt_shape = (btCollisionShape*) shape;
+        assert(bt_shape->getShapeType() == SPHERE_SHAPE_PROXYTYPE);
+        *radius = ((btSphereShape*) bt_shape)->getRadius();
+    }
+
+    void GetCollisionShapeCapsuleRadiusHeight3D(HCollisionShape3D shape, float* radius, float* half_height)
+    {
+        btCollisionShape* bt_shape = (btCollisionShape*) shape;
+        assert(bt_shape->getShapeType() == CAPSULE_SHAPE_PROXYTYPE);
+        btCapsuleShape* as_capsule = (btCapsuleShape*) bt_shape;
+        *radius                    = as_capsule->getRadius();
+        *half_height               = as_capsule->getHalfHeight();
+    }
+
+    void SetCollisionShapeRadius3D(HCollisionShape3D shape, float radius)
+    {
+        btCollisionShape* bt_shape = (btCollisionShape*) shape;
+        assert(bt_shape->getShapeType() == SPHERE_SHAPE_PROXYTYPE);
+        ((btSphereShape*) bt_shape)->setUnscaledRadius(radius);
+    }
+
+    void GetCollisionShapeHalfBoxExtents3D(HCollisionShape3D shape, float* xyz)
+    {
+        btCollisionShape* bt_shape = (btCollisionShape*) shape;
+        assert(bt_shape->getShapeType() == BOX_SHAPE_PROXYTYPE);
+        const btVector3& half_extents = ((btBoxShape*) bt_shape)->getHalfExtentsWithMargin();
+        xyz[0] = half_extents.getX();
+        xyz[1] = half_extents.getY();
+        xyz[2] = half_extents.getZ();
+    }
+
     void SetDebugCallbacks3D(HContext3D context, const DebugCallbacks& callbacks)
     {
         context->m_DebugCallbacks = callbacks;
+    }
+
+    void ReplaceShape3D(HCollisionObject3D object, HCollisionShape3D old_shape, HCollisionShape3D new_shape)
+    {
+        btCollisionObject* bt_object = GetCollisionObject(object);
+
+        btCollisionShape* bt_current_shape = bt_object->getCollisionShape();
+
+        if (bt_current_shape->isCompound())
+        {
+            btCompoundShape* compound_shape = (btCompoundShape*) bt_current_shape;
+            uint32_t n = compound_shape->getNumChildShapes();
+            for (uint32_t k = 0; k < n; ++k)
+            {
+                btCollisionShape* child = compound_shape->getChildShape(k);
+                if (child == old_shape)
+                {
+                    btTransform t = compound_shape->getChildTransform(k);
+                    compound_shape->removeChildShape(child);
+                    compound_shape->addChildShape(t, (btConvexShape*)new_shape);
+                    break;
+                }
+            }
+        }
+        else if (bt_current_shape == old_shape)
+        {
+            bt_object->setCollisionShape((btConvexShape*) new_shape);
+        }
     }
 
     void ReplaceShape3D(HContext3D context, HCollisionShape3D old_shape, HCollisionShape3D new_shape)
